HDDS-9706. Moved testOMRatisSnapshot test from TestOzoneManagerHAWithStoppedNodes to TestOzoneManagerHAWithAllRunning.
